Add rewards package and Calculator interface
