+++ icon/iiss/icstate/networkvalue.go
@@ -56,6 +56,7 @@ const (
 const (
 	IISSVersion2 int = iota + 2
 	IISSVersion3
+	IISSVersion4
 )
 
 const (
