+++ icon/iiss/module.go
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2023 ICON Foundation
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package iiss
+
+import (
+	"math/big"
+
+	"github.com/icon-project/goloop/common/db"
+	"github.com/icon-project/goloop/icon/iiss/icreward"
+)
+
+type Calculator interface {
+	Stop()
+	IsRunningFor(dbase db.Database, back, reward []byte) bool
+	WaitResult(blockHeight int64) error
+	TotalReward() *big.Int
+	Result() *icreward.Snapshot
+}
