+++ icon/iiss/rewards/iiss3_test.go
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 ICON Foundation
+ * Copyright 2023 ICON Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,19 +14,15 @@
  * limitations under the License.
  */
 
-package iiss
+package rewards
 
 import (
 	"fmt"
+	"github.com/stretchr/testify/assert"
 	"math/big"
 	"testing"
-	"time"
-
-	"github.com/stretchr/testify/assert"
 
 	"github.com/icon-project/goloop/common"
-	"github.com/icon-project/goloop/common/db"
-	"github.com/icon-project/goloop/common/errors"
 	"github.com/icon-project/goloop/common/log"
 	"github.com/icon-project/goloop/icon/icmodule"
 	"github.com/icon-project/goloop/icon/iiss/icreward"
@@ -34,88 +30,12 @@ import (
 	"github.com/icon-project/goloop/icon/iiss/icstate"
 )
 
-func MakeCalculator(database db.Database, back *icstage.Snapshot) *Calculator {
-	c := new(Calculator)
-	c.back = back
-	c.base = icreward.NewSnapshot(database, nil)
-	c.temp = c.base.NewState()
-	c.log = log.New()
-
-	return c
-}
-
-func TestCalculator_processClaim(t *testing.T) {
-	database := db.NewMapDB()
-	front := icstage.NewState(database)
-
-	addr1 := common.MustNewAddressFromString("hx1")
-	addr2 := common.MustNewAddressFromString("hx2")
-	v1 := int64(100)
-	v2 := int64(200)
-
-	type args struct {
-		addr  *common.Address
-		value *big.Int
-	}
-
-	tests := []struct {
-		name string
-		args args
-		want int64
-	}{
-		{
-			"Add Claim 100",
-			args{
-				addr1,
-				big.NewInt(v1),
-			},
-			v1,
-		},
-		{
-			"Add Claim 200",
-			args{
-				addr2,
-				big.NewInt(v2),
-			},
-			v2,
-		},
-	}
-
-	// initialize data
-	c := MakeCalculator(database, nil)
-	for _, tt := range tests {
-		args := tt.args
-		// temp IScore : args.value * 2
-		iScore := icreward.NewIScore(new(big.Int).Mul(args.value, big.NewInt(2)))
-		err := c.temp.SetIScore(args.addr, iScore)
-		assert.NoError(t, err)
-
-		// add Claim : args.value
-		_, err = front.AddIScoreClaim(args.addr, args.value)
-		assert.NoError(t, err)
-	}
-	c.back = front.GetSnapshot()
-
-	err := c.processClaim()
-	assert.NoError(t, err)
-
-	// check result
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			args := tt.args
-			iScore, err := c.temp.GetIScore(args.addr)
-			assert.NoError(t, err)
-			assert.Equal(t, 0, args.value.Cmp(iScore.Value()))
-		})
-	}
-}
-
 func TestCalculator_processBlockProduce(t *testing.T) {
 	addr0 := common.MustNewAddressFromString("hx0")
 	addr1 := common.MustNewAddressFromString("hx1")
 	addr2 := common.MustNewAddressFromString("hx2")
 	addr3 := common.MustNewAddressFromString("hx3")
-	variable := big.NewInt(int64(YearBlock * icmodule.IScoreICXRatio))
+	variable := big.NewInt(int64(icmodule.YearBlock * icmodule.IScoreICXRatio))
 	rewardGenerate := variable.Int64()
 	rewardValidate := variable.Int64()
 
@@ -239,13 +159,13 @@ func TestCalculator_varForVotedReward(t *testing.T) {
 				0,
 				100-1,
 				icmodule.RevisionIISS,
-				big.NewInt(YearBlock),
+				big.NewInt(icmodule.YearBlock),
 				big.NewInt(200),
 				22,
 				100,
 			),
 			//	multiplier = ((irep * MonthPerYear) / (YearBlock * 2)) * 100 * IScoreICXRatio
-			((YearBlock * MonthPerYear) / (YearBlock * 2)) * 100 * icmodule.IScoreICXRatio,
+			((icmodule.YearBlock * icmodule.MonthPerYear) / (icmodule.YearBlock * 2)) * 100 * icmodule.IScoreICXRatio,
 			1,
 		},
 		{
@@ -280,7 +200,7 @@ func TestCalculator_varForVotedReward(t *testing.T) {
 			),
 			// 	variable = iglobal * iprep * IScoreICXRatio / (100 * TermPeriod)
 			10000 * 50 * icmodule.IScoreICXRatio,
-			100 * MonthBlock,
+			100 * icmodule.MonthBlock,
 		},
 		{
 			"Global Version2 - disabled",
@@ -503,7 +423,7 @@ func TestVotedInfo_SortAndUpdateTotalBondedDelegationAndCalculateReward(t *testi
 	vInfo.UpdateTotalBondedDelegation()
 	assert.Equal(t, total, vInfo.TotalBondedDelegation().Int64())
 
-	variable := big.NewInt(YearBlock)
+	variable := big.NewInt(icmodule.YearBlock)
 	divider := big.NewInt(1)
 	period := 10000
 	bigIntPeriod := big.NewInt(int64(period))
@@ -554,7 +474,7 @@ func TestCalculator_varForVotingReward(t *testing.T) {
 					0,
 					100-1,
 					icmodule.RevisionIISS,
-					big.NewInt(MonthBlock),
+					big.NewInt(icmodule.MonthBlock),
 					big.NewInt(20000000),
 					22,
 					100,
@@ -562,8 +482,8 @@ func TestCalculator_varForVotingReward(t *testing.T) {
 				nil,
 			},
 			want{
-				RrepMultiplier * 20000000 * icmodule.IScoreICXRatio,
-				YearBlock * RrepDivider,
+				icmodule.RrepMultiplier * 20000000 * icmodule.IScoreICXRatio,
+				icmodule.YearBlock * icmodule.RrepDivider,
 			},
 		},
 		{
@@ -574,7 +494,7 @@ func TestCalculator_varForVotingReward(t *testing.T) {
 					0,
 					100-1,
 					icmodule.RevisionIISS,
-					big.NewInt(MonthBlock),
+					big.NewInt(icmodule.MonthBlock),
 					big.NewInt(0),
 					22,
 					100,
@@ -607,7 +527,7 @@ func TestCalculator_varForVotingReward(t *testing.T) {
 			// 	multiplier = iglobal * ivoter * IScoreICXRatio / (100 * TermPeriod, totalVotingAmount)
 			want{
 				10000 * 50 * icmodule.IScoreICXRatio,
-				100 * MonthBlock * 10,
+				100 * icmodule.MonthBlock * 10,
 			},
 		},
 		{
@@ -663,11 +583,11 @@ func TestCalculator_VotingReward(t *testing.T) {
 		string(addr3.Bytes()): {100, 200},
 	}
 
-	d0 := icstate.NewDelegation(addr1, big.NewInt(MinDelegation-1))
-	d1 := icstate.NewDelegation(addr1, big.NewInt(MinDelegation))
-	d2 := icstate.NewDelegation(addr2, big.NewInt(MinDelegation))
-	d3 := icstate.NewDelegation(addr3, big.NewInt(MinDelegation))
-	d4 := icstate.NewDelegation(addr4, big.NewInt(MinDelegation))
+	d0 := icstate.NewDelegation(addr1, big.NewInt(icmodule.MinDelegation-1))
+	d1 := icstate.NewDelegation(addr1, big.NewInt(icmodule.MinDelegation))
+	d2 := icstate.NewDelegation(addr2, big.NewInt(icmodule.MinDelegation))
+	d3 := icstate.NewDelegation(addr3, big.NewInt(icmodule.MinDelegation))
+	d4 := icstate.NewDelegation(addr4, big.NewInt(icmodule.MinDelegation))
 	type args struct {
 		iissVersion int
 		multiplier  int
@@ -787,54 +707,4 @@ func TestCalculator_VotingReward(t *testing.T) {
 			assert.Equal(t, tt.want, reward.Int64())
 		})
 	}
-}
-
-func TestCalculator_WaitResult(t *testing.T) {
-	c := &Calculator{
-		startHeight: InitBlockHeight,
-	}
-	err := c.WaitResult(1234)
-	assert.NoError(t, err)
-
-	c = &Calculator{
-		startHeight: 3414,
-	}
-	err = c.WaitResult(1234)
-	assert.Error(t, err)
-
-	toTC := make(chan string, 2)
-
-	go func() {
-		err := c.WaitResult(3414)
-		assert.True(t, err == errors.ErrInvalidState)
-		toTC <- "done"
-	}()
-	time.Sleep(time.Millisecond * 10)
-
-	c.setResult(nil, errors.ErrInvalidState)
-	assert.Equal(t, "done", <-toTC)
-
-	c = &Calculator{
-		startHeight: 3414,
-	}
-	go func() {
-		err := c.WaitResult(3414)
-		assert.NoError(t, err)
-		toTC <- "done"
-	}()
-	go func() {
-		err := c.WaitResult(3414)
-		assert.NoError(t, err)
-		toTC <- "done"
-	}()
-	time.Sleep(time.Millisecond * 20)
-
-	mdb := db.NewMapDB()
-	rss := icreward.NewSnapshot(mdb, nil)
-	c.setResult(rss, nil)
-
-	assert.Equal(t, "done", <-toTC)
-	assert.Equal(t, "done", <-toTC)
-
-	assert.True(t, c.Result() == rss)
 }
